refactor: brandstory page 컴포넌트화
